Fixed isLoggedIn (again)
Ever wanted to do...anything in the Toonbook Buddy window, but got
halted every couple of seconds? I did. I got ticked off.

Instead of returning a true or false value, isLoggedIn() now takes 2
parameters, a function for when the user is logged in, and an optional
function for when the user is logged out.